NumberOfUniqueNumberInAnArray
Write a Program to find out the sum of unique numbers of an array using the methods
2
Write the method with the following specifications:
Name of method sumOfArray()
Arguments: one argument of type integer array
Return Type: an integer value
Specifications: The value returned by the method sumOfArray() is determined by the following rules:
Array must not be null or length must not be 0. If yes, then return -1
Array must not contain negative values. If yes, then return -2
Input: {1, 2, 3, 2, 4, 1}
Output: 7