LambdaKeyValueExample
In the above code, we first create a HashMap and add some key-value pairs to it. Then, we demonstrate two ways to work with the key-value pairs using lambda expressions.

In the first part, we use the forEach method of the Map interface to iterate over the key-value pairs. The lambda expression (key, value) -> System.out.println("Key: " + key + ", Value: " + value) is used to define the action to be performed for each pair, which in this case is printing the key and value.

In the second part, we use the stream() method to convert the map to a stream, and then we filter the key-value pairs based on a condition using the filter method. The lambda expression entry -> entry.getValue() > 2 specifies the condition that only key-value pairs with values greater than 2 will be included. Finally, we use the forEach method to print the filtered key-value pairs.