HarshadNumber
GCD and LCM are two commonly used terms in mathematics, specifically when dealing with multiple numbers. 

GCD stands for "Greatest Common Divisor," also known as the highest common factor. It refers to the largest positive integer that divides two or more numbers without leaving a remainder. In other words, it is the largest number that divides all the given numbers evenly. The GCD is often used to simplify fractions or find common factors between numbers.

LCM stands for "Least Common Multiple." It refers to the smallest positive integer that is divisible by two or more numbers without leaving a remainder. In other words, it is the smallest number that is a multiple of all the given numbers. The LCM is often used to find a common multiple between numbers or to solve problems involving periodic events.

Both GCD and LCM are important concepts in number theory and have various applications in mathematics, computer science, and real-life problem-solving. They are used to analyze and manipulate numbers, simplify calculations, and solve problems involving multiple quantities.