SpyNumber
A Spy number is a concept in recreational mathematics. It is a positive integer that satisfies a particular property related to its digits.

Here's how a Spy number is determined:

1. Take a positive integer and find the sum of its digits.
2. Find the product of the digits.
3. Subtract the product from the sum.
4. If the result of the subtraction is equal to the original number, then it is called a Spy number.

For example, let's take the number 112:

Sum of digits = 1 + 1 + 2 = 4
Product of digits = 1 * 1 * 2 = 2
Difference: 4 - 2 = 2

Since the difference (2) is not equal to the original number (112), it is not a Spy number.

On the other hand, if we take the number 1124:

Sum of digits = 1 + 1 + 2 + 4 = 8
Product of digits = 1 * 1 * 2 * 4 = 8
Difference: 8 - 8 = 0

In this case, the difference (0) is equal to the original number (1124), so it is considered a Spy number.

Spy numbers are interesting from a recreational mathematics perspective, but they don't have significant applications in other fields.