PrimeNumber
A prime number is a natural number greater than 1 that has no positive divisors other than 1 and itself. In other words, a prime number is only divisible by 1 and itself, and it does not have any other factors.

For example, let's consider the number 7. The factors of 7 are 1 and 7. Since there are no other factors, 7 is a prime number.

Here are a few more examples of prime numbers:

- 2: The only factors of 2 are 1 and 2, making it a prime number.
- 13: The only factors of 13 are 1 and 13.
- 23: The only factors of 23 are 1 and 23.

On the other hand, numbers that have factors other than 1 and themselves are called composite numbers. For example, the number 4 has factors 1, 2, and 4, so it is not a prime number.

Prime numbers play a crucial role in number theory and have many applications in various fields, such as cryptography, data encryption, and algorithms.